refactor: flex 컴포넌트 타입 분리
