+++ src/features/common/components/Flex/Flex.tsx
@@ -1,10 +1,37 @@
+import { flexStyle } from './style.css'
+import { type CommonProps } from '@/types/common'
+import { Slot } from '@radix-ui/react-slot'
+import { getGapSizeFromNumber } from './utils'
 import { type HTMLAttributes } from 'react'
+import clsx from 'clsx'
+import { type FlexRecipeProps } from './type'
 
-type FlexProps = HTMLAttributes<HTMLDivElement> & {
-  direction?: 'row' | 'column'
-  justify?: 'start' | 'end' | 'center' | 'between' | 'around'
-  align?: 'start' | 'end' | 'center' | 'baseline' | 'stretch'
-}
-export function Flex({ children, ...props }: FlexProps) {
-  return <div {...props}>{children}</div>
+export type FlexProps = FlexRecipeProps &
+  CommonProps & {
+    gap?: number
+    asChild?: boolean
+  } & HTMLAttributes<HTMLDivElement>
+
+export const Flex = (props: FlexProps) => {
+  const {
+    direction,
+    align,
+    justify,
+    wrap,
+    asChild,
+    gap,
+    children,
+    className: classNameFromProps,
+    style: styleFromProps,
+    ...restProps
+  } = props
+  const Component = asChild ? Slot : 'div'
+  return (
+    <Component
+      style={{ gap: gap ? getGapSizeFromNumber(gap) : undefined, ...styleFromProps }}
+      className={clsx(flexStyle({ direction, align, justify, wrap }), classNameFromProps)}
+      {...restProps}>
+      {children}
+    </Component>
+  )
 }
