+++ src/features/common/components/Flex/type.ts
@@ -1,0 +1,4 @@
+import { type RecipeVariants } from '@vanilla-extract/recipes'
+import { type flexStyle } from './style.css'
+
+export type FlexRecipeProps = RecipeVariants<typeof flexStyle>
